notification-service/requirements.txt
fastapi
uvicorn
pydantic
pymongo
certifi
pika
python-dotenv